src/main/java/BankAccount.java

public class BankAccount {

    private String firstName;
    private String lastName;
    private String dateOfBirth;
    private int accountNumber;
    private int balance;

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setDateOfBirth(String dateOfBirth) {
        this.dateOfBirth = dateOfBirth;
    }

    public String getDateOfBirth() {
        return dateOfBirth;
    }

    public void setAccountNumber(int accountNumber) {
        this.accountNumber = accountNumber;
    }

    public int getAccountNumber() {
        return accountNumber;
    }

    public void setBalance(int balance) {
        this.balance = balance;
    }

    public int getBalance() {
        return balance ;
    }

    public void deposit(int depositAmount) {
        setBalance(getBalance() + depositAmount);
    }
    public void withdrawal(int withdrawalAmount){
        setBalance(getBalance() - withdrawalAmount);
    }

    public void payInterest(int  interestAmount){
        setBalance(getBalance() * interestAmount);
    }

}



